fix: Rename GetProperty to getSerializedProperty
This commit is part of a rebase to fix commit messages to new standards and cherry-picking commit 921e4f0.

Date: Tue Dec 19 22:19:40 2023 -0500

Rename `GetProperty` to `getSerializedProperty`
Add the ability to get the targeted object that a Serialized Property belongs to, as well as the ability to get the fields, find a field, and get the value of a field.

Changes:
 - modified:   Editor/Attributes/Properties/ShowIfBoolPropertyDrawer.cs
 - modified:   Editor/Attributes/Properties/ShowIfEnumPropertyDrawer.cs
 - modified:   Editor/SerializedPropertyExtensions.cs