feat(Settings): Add Menu for Enhanced Navigation and Display Order
This commit introduces a new feature as part of a rebase to update commit messages to new standards and cherry-picking commit 01bcf31.

Date: Mon Feb 5 21:52:54 2024 -0500

This feature adds a menu to enhance navigation to the relevant project settings and user preferences.

Additionally, it updates the CoreFrameworkProjectSettingsProvider to display before other project settings in the Core Path.